apps/promptshield-enterprise-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptshield-enterprise-api"
version = "0.1.0"
description = "PromptShield Enterprise API - centralized governance control plane"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "promptshield-core",
    "promptshield-config",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]

[tool.hatch.build.targets.wheel]
packages = ["promptshield_enterprise"]
